Fixed the issue of SRV High Beam overflowing the flags and activating everything

+++ src/typedefs.hpp
@@ -2,8 +2,8 @@
 #define EDFUEL_TYPEDEFS_INCLUDED
 
 typedef struct {
-  uint32_t flags;
-  uint32_t flags2;
+  uint64_t flags;
+  uint64_t flags2;
   unsigned long cmdtime;
 } flags_t;
